requirements.txt
python-telegram-bot==20.0
selenium==4.21.0
webdriver-manager==4.0.0  # Tự động quản lý ChromeDriver
requests==2.31.0
beautifulsoup4==4.12.0
lxml==5.2.1  # Parser nhanh cho BeautifulSoup